Correction: write output within loop and calculate query region properly
1) The write_div function is removed and we put the writing to_csv
again into the loop in the function divergence().

2) Also, the query regions were calculated wrongly (formerly called
windows): first, it started from 1 not 0, second, the boundaries of
adjacent regions overlapped such that some positions were occasionally
queried twice (-> duplicates).

3) Goodie: during calculation the step size (based on LOAD) and the
currently queried region are printed to stdout, which is useful for
monitoring. Also, if step size is maximal for the contig this is said
in the output to indicate that it does not make sense to increase LOAD
anymore.