WIP: add jsd_average and for js_divergence
- jsd_average works
- js_divergence works in principle:
  TODO:
  - pool.map jsd_average does not work
  - output format not correct

**BUG**
During debugging, I realized that the average JSD of a lower level
is highe than the total JSD; this is not possible; a check showed that
function jsd_is() assigned for some loci an unrealistically high value
(a subset with a higher meth. level at that locus had a lower JSD than
a subset with a lower meth. level); this is inconsistent and seems to be
a BUG in jsd_is!